requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers
psycopg2-binary==2.9.10
python-dotenv==1.1.0
dj-database-url==0.5.0
celery
redis
telebot